Add configuration, webhook processing, and WATI service integration with event templates
